requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements/requirements.in -o ./requirements/requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.49.0
    # via -r ./requirements/requirements.in
anyio==4.9.0
    # via
    #   anthropic
    #   google-genai
    #   httpx
    #   openai
    #   starlette
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
comm==0.2.2
    # via ipykernel
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via
    #   anthropic
    #   openai
executing==2.2.0
    # via stack-data
fastapi==0.115.12
    # via -r ./requirements/requirements.in
google-auth==2.38.0
    # via google-genai
google-genai==1.10.0
    # via -r ./requirements/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   google-genai
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
ipykernel==6.29.5
    # via -r ./requirements/requirements.in
ipython==9.1.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jiter==0.9.0
    # via
    #   anthropic
    #   openai
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mouseinfo==0.1.3
    # via pyautogui
nest-asyncio==1.6.0
    # via ipykernel
numpy==2.2.4
    # via pandas
openai==1.72.0
    # via -r ./requirements/requirements.in
packaging==24.2
    # via ipykernel
pandas==2.2.3
    # via -r ./requirements/requirements.in
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.1.0
    # via -r ./requirements/requirements.in
platformdirs==4.3.7
    # via jupyter-core
prompt-toolkit==3.0.50
    # via ipython
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyautogui==0.9.54
    # via -r ./requirements/requirements.in
pydantic==2.11.3
    # via
    #   -r ./requirements/requirements.in
    #   anthropic
    #   fastapi
    #   google-genai
    #   openai
pydantic-core==2.33.1
    # via pydantic
pygetwindow==0.0.9
    # via pyautogui
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
pymsgbox==1.0.9
    # via pyautogui
pyobjc-core==11.0
    # via
    #   pyautogui
    #   pyobjc-framework-cocoa
    #   pyobjc-framework-quartz
pyobjc-framework-cocoa==11.0
    # via pyobjc-framework-quartz
pyobjc-framework-quartz==11.0
    # via pyautogui
pyperclip==1.9.0
    # via
    #   -r ./requirements/requirements.in
    #   mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   pandas
python-dotenv==1.1.0
    # via -r ./requirements/requirements.in
pytweening==1.2.0
    # via pyautogui
pytz==2025.2
    # via pandas
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.3
    # via google-genai
rsa==4.9
    # via google-auth
rubicon-objc==0.5.0
    # via mouseinfo
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
stack-data==0.6.3
    # via ipython
starlette==0.46.1
    # via fastapi
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.13.2
    # via
    #   anthropic
    #   anyio
    #   fastapi
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.0
    # via -r ./requirements/requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via google-genai
